Move all modals and gameplay side panel components to new 'features' directory
