+++ frontend/src/features/modals/gameOverModals/LocalModal.tsx
