+++ frontend/src/pages/protected/PassAndPlay.tsx
@@ -11,7 +11,7 @@ import "../../styles/components/chessboard/board-actions.scss";
 import "../../styles/pages/pass-and-play.scss";
 
 import Chessboard from "../../components/global/chessboards/Chessboard.tsx";
-import GameplaySettings from "../../components/global/modals/GameplaySettings.tsx";
+import GameplaySettings from "../../features/modals/GameplaySettings.tsx";
 import ModalWrapper from "../../components/global/wrappers/ModalWrapper.tsx";
 import useGameplaySettings from "../../hooks/useGameplaySettings.ts";
 import { ParsedFENString } from "../../types/gameLogic.ts";
